NA OFI: pass down na flags for firewall support in prov/tcp
This PR allows to pass down a flag to indicate if client bulk address is
behind firewall by using address deserialization callback functions.

This flag will lead to FI_FIREWALL_ADDR flag to be set in OFI.

+++ src/mercury_proc_bulk.c
@@ -36,7 +36,7 @@ hg_proc_hg_bulk_t(hg_proc_t proc, void *data)
 
     switch (hg_proc_get_op(proc)) {
         case HG_ENCODE: {
-            uint8_t flags = 0;
+            uint32_t flags = 0;
             bool try_eager = false; /* Flag will not be set if bulk handle does
                                        not support it */
